refactor: extract ShinyGradientButton
